Refactor code to remove switch case statement
The switch case statement was inflexible and a terrible design. I replaced the checkandSetValue method with a setValue method that accepts a function pointer to a check function. This is necessary because the check function is different depending on the property.

The previous design was also exposing a pointer to the private MeasConfigType struct which completely defeated the point of the private struct. This has been replaced by private properties for each value in the MeasConfigType struct which are configured using setter and getter methods. This entire configuration class can now be passed to the constructor of the Spectrometer class which will call all of the getter methods and configure a MeasConfigType struct internally.