Add validators to constructor
Create a map whose keys are the names of the parameters as strings and whose values are pointers to the MeasConfigType values. The constructor takes a configuration vector whose values are spec_config_param vectors. The constructor loops over the values in this vector, validates them and then assigns them to the MeasConfigType spec_config member.